Remove all XML formatting on RequestBlock WriteXml()
I was told to make this change and PR by Intacct's Director of Technical Support to deal with a performance issue we are seeing in API usage. The Engineering team wanted to make sure that the API request body is sent as one long unformatted string to reduce API request size to help keep requests under the 500k character limit.